+++ clash-lib/src/CLaSH/Driver.hs
@@ -25,6 +25,7 @@ import           CLaSH.Core.Term                  (Term)
 import           CLaSH.Core.Type                  (Type)
 import           CLaSH.Core.TyCon                 (TyCon, TyConName)
 import           CLaSH.Driver.TestbenchGen
+import           CLaSH.Driver.TopWrapper
 import           CLaSH.Driver.Types
 import           CLaSH.Netlist                    (genNetlist)
 import           CLaSH.Netlist.Types              (Component (..), HWType)
@@ -42,9 +43,10 @@ generateHDL :: forall backend . Backend backend
             -> HashMap TyConName TyCon -- ^ TyCon cache
             -> (HashMap TyConName TyCon -> Type -> Maybe (Either String HWType)) -- ^ Hardcoded 'Type' -> 'HWType' translator
             -> (HashMap TyConName TyCon -> Term -> Term) -- ^ Hardcoded evaluator (delta-reduction)
+            -> Maybe TopEntity
             -> DebugLevel -- ^ Debug information level for the normalization process
             -> IO ()
-generateHDL bindingsMap hdlState primMap tcm typeTrans eval dbgLevel = do
+generateHDL bindingsMap hdlState primMap tcm typeTrans eval teM dbgLevel = do
   start <- Clock.getCurrentTime
   prepTime <- start `deepseq` bindingsMap `deepseq` tcm `deepseq` Clock.getCurrentTime
   let prepStartDiff = Clock.diffUTCTime prepTime start
@@ -105,7 +107,8 @@ generateHDL bindingsMap hdlState primMap tcm typeTrans eval dbgLevel = do
       putStrLn $ "Testbench generation took " ++ show netTBDiff
 
       let hdlState' = fromMaybe (initBackend :: backend) hdlState
-          hdlDocs = createHDL hdlState' (netlist ++ testBench)
+          topWrapper = mkTopWrapper primMap teM topComponent
+          hdlDocs = createHDL hdlState' (topWrapper : netlist ++ testBench)
           dir = concat [ "./" ++ CLaSH.Backend.name hdlState' ++ "/"
                        , takeWhile (/= '.') (name2String $ fst topEntity)
                        , "/"
@@ -122,8 +125,8 @@ generateHDL bindingsMap hdlState primMap tcm typeTrans eval dbgLevel = do
 
 -- | Pretty print Components to HDL Documents
 createHDL :: Backend backend
-           => backend
-           -> [Component]
+           => backend     -- ^ Backend
+           -> [Component] -- ^ List of components
            -> [(String,Doc)]
 createHDL backend components = flip evalState backend $ do
   (hdlNms,hdlDocs) <- unzip <$> mapM genHDL components