+++ clash-vhdl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`clash-vhdl`](http://hackage.haskell.org/package/clash-vhdl) package
 
+## 0.5.3
+* New features:
+  * Support wrapper generation
+
 ## 0.5.2 *April 24th 2015*
 * Fixes bugs:
   * Fix bug where not enough array type definitions were created